+++ examples/dartvel_example/.dartvel/ssg_builder.dart
@@ -2,15 +2,11 @@
 import 'dart:convert';
 import 'dart:io';
 import 'package:dartvel_flutter/dartvel_flutter.dart';
-import 'package:dartvel_example/pages/about.page.dart' deferred as p0;
-import 'package:dartvel_example/dartvel_client/dartvel_client.dart';
-import 'package:flutter/material.dart';
-import 'package:dartvel_example/pages/blog/[id].page.dart' deferred as p1;
-import 'package:dartvel_example/pages/index.page.dart' deferred as p2;
-import 'package:dartvel_example/pages/next_shift.page.dart' deferred as p3;
-import 'package:dartvel_example/widgets/next_shift.dart';
-import 'package:dartvel_example/pages/pricing.dart' deferred as p4;
-import 'package:flutter/widgets.dart';
+import 'package:dartvel_example/dartvel_client/pages/about_page.g.dart' deferred as p0;
+import 'package:dartvel_example/dartvel_client/pages/blog_id_page.g.dart' deferred as p1;
+import 'package:dartvel_example/dartvel_client/pages/index_page.g.dart' deferred as p2;
+import 'package:dartvel_example/dartvel_client/pages/next_shift_page.g.dart' deferred as p3;
+import 'package:dartvel_example/dartvel_client/pages/pricing_page.g.dart' deferred as p4;
 void main() async {
   final outDir = Directory('build/web/_ssg');
   if (!outDir.existsSync()) outDir.createSync(recursive: true);